+++ src/features/artist/components/ArtistHeader.tsx
