+++ src/features/artist/components/ArtistInfo.tsx
