+++ src/features/artist/components/ArtistSkeleton.tsx
